migrate to @bubble-protocol/client v0.1.0 (bubble managers)
