+++ src/model/Chat.js
@@ -1,0 +1,280 @@
+import { BubbleManager, ManagedBubble, bubbleProviders, assert, toFileId } from "@bubble-protocol/client";
+import { ContentId } from "@bubble-protocol/core";
+import localStorage from "./utils/LocalStorage";
+import { EventManager } from "./utils/EventManager";
+import { stateManager } from "../state-context";
+import { User } from "./User";
+
+const CONTENT = {
+  metadataFile: toFileId(101),
+  textChat: toFileId(1),
+}
+
+const DEFAULT_METADATA = {
+  title: undefined,
+  icon: undefined,
+  members: []
+}
+
+const STATE = {
+  invalid: 'invalid',
+  connecting: 'connecting',
+  open: 'open',
+  reconnecting: 'reconnecting',
+  failed: 'failed'
+}
+
+export class Chat extends ManagedBubble {
+
+  state = STATE.uninitialised;
+  listeners = new EventManager(['new-message-notification', 'unread-change']);
+
+  user;
+  metadata;
+  messages = [];
+  lastModTime = 0;
+  
+  constructor(bubbleId, myId, deviceKey, encryptionPolicy, contentManager) {
+    assert.isObject(myId, 'myId');
+    assert.isObject(deviceKey, 'deviceKey');
+    bubbleId = new ContentId(bubbleId);
+
+    // construct the provider
+    const provider = new bubbleProviders.WebsocketBubbleProvider(bubbleId.provider, {sendTimeout: 10000});
+
+    // construct the bubble
+    super(bubbleId, provider, deviceKey.signFunction, encryptionPolicy, contentManager || new BubbleManager());
+    this.id = bubbleId.chain+'-'+bubbleId.contract;
+    this.myId = myId;
+
+    // register state variables
+    stateManager.register(this.id+'-connection-state', this.provider.state);
+    stateManager.register(this.id+'-metadata', {bubbleId: bubbleId, ...DEFAULT_METADATA});
+    stateManager.register(this.id+'-messages', []);
+    stateManager.register(this.id+'-unread', 0);
+
+    // setup listeners
+    provider.on('event', (_, state) => stateManager.dispatch(this.id+'-connection-state', state))
+    provider.on('reconnect', this._handleReconnect.bind(this));
+    provider.on('error', event => console.warn(this.id, 'websocket error:', event));
+    this.on = this.listeners.on.bind(this.listeners);
+    this.off = this.listeners.off.bind(this.listeners);
+  }
+
+  create({metadata={}, options}) {
+    this._validateMetadata(metadata);
+    return super.create(options)
+      .then(() => {
+        return Promise.all([
+          this._saveMetadata(metadata, options),
+          this.mkdir(CONTENT.textChat, options),
+        ])
+      })
+      .then(() => {
+        return this._subscribeToContent(false, false, options);
+      })
+      .then(() => {
+        this.state = STATE.initialised;
+        return this.contentId;
+      })
+  }
+
+  initialise(options) {
+    return this._loadMessages(this.id)
+      .then(() => super.initialise(options))
+      .then(() => {
+        return this._subscribeToContent(true, true, options);
+      })
+      .then(() => {
+        this.state = STATE.initialised;
+        return this.metadata;
+      })
+  }
+
+  join() {
+    console.trace('joining conversation bubble');
+    return super.initialise()
+      .then(() => {
+        return this._subscribeToContent(true, true);
+      })
+      .then(() => {
+        return this.metadata;
+      })
+  }
+
+  postMessage(message) {
+    console.trace('post message', this.id, message);
+    if (message.id === undefined) message.id = Date.now() + Math.floor(Math.random() * Math.pow(10, 6));
+    this.write(CONTENT.textChat + '/' + message.id, JSON.stringify(message))
+      .then(() => {
+        message.created = Date.now();
+        message.modified = message.created;
+        this._setMessage(message);
+      })
+    message.pending = true;
+    message.created = Date.now();
+    message.modified = message.created;
+    this._setMessage(message);
+    return Promise.resolve();
+  }
+
+  addUser(publicKey, options) {
+    if (!this.contentManager || !assert.isFunction(this.contentManager.addUser)) throw new Error('not a multi-user chat');
+    return this.contentManager.addUser(this, publicKey, {...options, userMetadata: this.userMetadata});
+  }
+
+  setMetadata(metadata) {
+    this._validateMetadata(metadata);
+    this._saveMetadata(metadata);
+  }
+
+  setReadTime(time) {
+    this.lastRead = time;
+    this._updateUnread();
+  }
+
+  getMessages() {
+    return this.messages || [];
+  }
+
+  serialize() {
+    return {id: this.id, bubbleId: this.bubbleId.toString(), metadata: this.metadata}
+  }
+
+  deserialize(data) {
+    assert.isObject(data, 'data');
+    assert.isString(data.id, 'data');
+    assert.isString(data.bubbleId, 'data');
+    if (data.id !== this.id) throw new Error('wrong id in serialized data');
+    if (data.bubbleId !== this.bubbleId.toString()) throw new Error('wrong bubbleId in serialized data');
+    this._validateMetadata(data.metadata);
+    this.metadata = data.metadata;
+  }
+
+  isValid() {
+    return this.state !== STATE.invalid;
+  }
+
+  close() {
+    return this.provider.close();
+  }
+
+  _validateMetadata(metadata) {
+    assert.isObject(metadata, 'metadata');
+    assert.isString(metadata.title, 'title');
+    assert.isArray(metadata.members, 'members');
+  }
+
+  _saveMetadata(metadata, options) {
+    return this.write(CONTENT.metadataFile, JSON.stringify(metadata), options);
+  }
+
+  _handleMessageNotification(notification) {
+    console.trace('msg rxd', notification);
+    if (assert.isArray(notification.data)) {
+      notification.data.forEach(msg => this._readMessage(msg));
+    }
+  }
+
+  _handleMetadataChange(notification) { 
+    const metadata = JSON.parse(notification.data);
+    console.debug('new metadata', metadata)
+    this.metadata = {...DEFAULT_METADATA, ...metadata, bubbleId: this.contentId};
+    stateManager.dispatch(this.id+'-metadata', this.metadata);
+  }
+
+  _readMessage(messageDetails, attempts=5) {
+    console.trace('reading message', messageDetails);
+    return this.read(messageDetails.name)
+      .then(messageJson => {
+        try { console.debug('_readMessage', messageJson);
+          const message = JSON.parse(messageJson);
+          message.created = messageDetails.created;
+          message.modified = messageDetails.modified;
+          if (assert.isString(message.from)) this._setMessage(message);
+          else console.warn('invalid message rxd', message);
+        }
+        catch(error) {
+          console.warn(this.id, 'message parse error', error);
+        }
+      })
+      .catch(error => {
+        console.warn(this.id, 'message read error', error.code, error.message);
+        if (error.code === -32005) { 
+          // bubble server internal error.  Try again in 1s.
+          if (attempts <= 1) console.warn(this.id, 'abandoning message read after 5 failed attempts');
+          else {
+            console.trace(this.id, 'trying again in 1s')
+            setTimeout(() => this._readMessage(messageDetails, --attempts), 1000);
+          }
+        }
+      })
+  }
+
+  _setMessage(message) {
+    message.from = new User(message.from)
+    const index = this.messages.findIndex(m => m.id === message.id);
+    if (index >=0) this.messages[index] = message;
+    else this.messages.push(message);
+    this.messages.sort((a,b) => a.created - b.created);
+    this.messages = [...this.messages];  // mutate to trigger any react hooks
+    localStorage.writeMessage(message);
+    stateManager.dispatch(this.id+'-messages', this.messages);
+    this.listeners.notifyListeners('new-message-notification');
+  }
+
+  _loadMessages(conversationId) {
+    return localStorage.queryMessagesByConversation(conversationId)
+      .then(messages => { console.debug('loaded messages', messages);
+        messages.sort((a,b) => a.created - b.created);
+        messages.forEach(m => {m.from = new User(m.from)});
+        this.messages = messages;
+        this.lastModTime = messages.reduce((time, m) => {return m.modified > time ? m.modified : time}, 0);
+        stateManager.dispatch(this.id+'-messages', this.messages);
+      })
+  }
+
+  _handleReconnect() {
+    this._subscribeToContent(true, true)
+      .then(() => {
+        this.listeners.notifyListeners('state-change', this.getConnectionState());
+        this.listeners.notifyListeners('available', true)
+      })
+      .catch(error => {
+        console.warn('bubble failed to resubscribe after reconnection', error);
+        this.provider.close();
+      });
+  }
+
+  _subscribeToContent(metadataRead, textChatList, options={}) {
+    console.trace('subscribing to content');
+    return Promise.all([
+      this.subscribe(CONTENT.metadataFile, this._handleMetadataChange.bind(this), {...options, read: metadataRead}),
+      this.subscribe(CONTENT.textChat, this._handleMessageNotification.bind(this), {...options, since: textChatList ? this.lastModTime : undefined}),
+    ])
+    .then(subscriptions => { console.debug('subscriptions', subscriptions)
+      if(metadataRead) this._handleMetadataChange(subscriptions[0]);
+      if(textChatList) this._handleMessageNotification(subscriptions[1]);
+    })
+  }
+
+  _updateUnread() {
+    this.unreadMsgs = this._countUnreadMsgs();
+    console.debug(this.id+' unread', this.unreadMsgs)
+    stateManager.dispatch(this.id+'-unread', this.unreadMsgs);
+    this.listeners.notifyListeners('unread-change');
+  }
+
+  _countUnreadMsgs() {
+    let count = 0;
+    let index = this.messages.length-1;
+    while (index >= 0 && this.messages[index].modified > this.lastRead) {
+      if (this.messages[index].from.id !== this.myId.id) count++;
+      index--;
+    }
+    return count;  
+  }
+
+
+}
+
